Add final answers to state
Setting a final answer in state (eg. in a router, or in a lifecycle)
forces the network to terminate.

This lets us have a single, unified way to extract results from an
agentic loop.

Note that the final answer can be of any shape, and is currently typed
as `unknown`.