fix(tools): OpenAI nullable tool parameters support
As stated in OpenAI docs, in strict mode, all tool parameters must
be required (https://platform.openai.com/docs/guides/structured-outputs?context=with_parse#all-fields-must-be-required).
However, it is possible to emulate "nullable" field with a JSONSchema union type
which is not supported by standard libraries.
OpenAI has a vendored version of `zod-to-json-schema` to support this.
Good news is that `zod-to-json-schema` backported OpenAI vendored version in their lib in last December 2024.