Bug: State.clone is always resetting existing state (#149)
Refactoring gone wrong here it seems. The constructor args for `State` are `({data, messages})` (it was `(data)` before that), but it's been passed `state.data` object directly.

This causes issues when `NetworkRun` is created with a cloned state – it doesn't include any of the data from the state that exists in the `Network`.